Updated Header
Created a new Header component, and have moved it to being a child of the Layout component so that it renders on all pages using the Layout component instead.

Still needs "active indicators" tied in, and the "dark" or "light" version to be rendered depending on which page the router is on.